Fixed a bug that would cause multiple instreams to fail
When the event handler is removed on stop, it is not correctly re-added on start (only if _timer was not set at all).
This causes multiple inStreams to fail, i.e. the first one would play just fine, but after that the time would never execute the "onTimer" method; thus no other inStream would ever play.